Make fields that might not be published optional
While we don't want content-tree to model where the data comes from, we
have to handle the fact that when we publish the content-tree to the
content API, some fields won't be part of that tree (e.g. the tweet
HTML).

To solve this, we can make these properties optiona