add tables to the tree
The property names could use some work.
There is no concept of header rows, or rows at all.
The belief is that with the `rows` and `columns` properties you should
be able to work all that out yourself. the header row is always
the first $rows number of cells.

in spark there is always 1 header row, and the rows can never have
variable numbers of clls.