add initial content-tree-to-text implementation Based on the transformer from here:
https://github.com/Financial-Times/ccf-content-rw-elasticsearch/blob/b01e6aab54d53167ff5e8b59c25171a023978c60/pkg/html/transformer.go

There are a few test case input->expected output in the root tests/ folder that could be used by implementations in other languages, so we know we're all doing it the same way. We can add more tests as we like. I've just done an old methode article and a Spark kitchen sink so far.

folders have been added for test cases have also been added for a future content-tree->external-content bodyxml converter (will be used for external consumers in 2024 and beyond), and a bodyxml->content-tree (for a future migration tool 👀 for our old articles).

This is just a first pass, can probably be improved, mostly here to register the idea and move towards the future