Add an initial clip-set definition
This should be right for how they work right now, but I'm going to suggest we don't merge this.

Currently `"clip-set"` is very different from the other things in content-tree (property names and data structure) so i think it would behoove us to get it more in line before we make this final, that'll involve some work with customer products in their cp-content-pipeline repo. But i've been informed that they'd be well up for it.