Add Go implementation of the content tree full schema
The Go structs allow representation of full content tree
and converting a JSON content tree to Go objects (unmarshalling).
Add Go mod definition in the main dir of the project.

This implementation returns an error during unmarshalling if the
tree does not adhere to the JSON schema, including when the node
type is "undefined".