Use /internalcontent kitchen sink body, instead of Spark
The test fixture we were using came from Spark - however C&M do various
transformations which mean the bodyXML we publish from Spark isn't
necessarily the same as the bodyXML that a user would read from the API.

For the purpose of this transformer, we are interested in the latter, so
this change updates the test fixture file to be the published version of
the bodyXML.

I have also taken the expected JSON from the bodyTree produced by
cp-content-pipeline-api (removing the "data" blocks that contain
referenceIds). There may be other places where the CP-generated tree is
not actually valid, but will deal with that as we go along.